Add files via upload
in the UCI Machine Learning Repository. This dataset consists of 21,263 superconducting materials where each of them has a total of 81 extracted features (like atomic mass, valence, electronegativity, etc). The primary goal will be to make prediction models to try and estimate the critical temperature (Tc) of superconductors based on the features. We will use regression methods and aims to find the most important predictors to Tc and examine different machine learning algorithms on predicting Tc. The results will build an understanding of superconducting materials as well as help discover new superconductors with convenient properties.